requirements.txt
configparser==3.5.0
Django==1.11.4
django-cors-headers==2.1.0
djangorestframework==3.6.3
djangorestframework-jsonapi==2.0.0b2
enum34==1.1.6
flake8==3.4.1
inflection==0.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2017.2
